+++ packages/onionprobe/tls.py
@@ -26,14 +26,23 @@ except ImportError:
 
 import ssl
 
+from cryptography.x509              import load_pem_x509_certificate
+from cryptography.hazmat.primitives import hashes
+
 class OnionprobeTLS:
     """
     Onionprobe class with TLS methods.
     """
 
-    def get_certificate(self, endpoint, config, attempt = 1):
+    def query_tls(self, endpoint, config, attempt = 1):
         """
-        Gets the endpoint's TLS certificate.
+        Tries a TLS connection to the endpoint and update metrics when needed.
+
+        This method does not make any certificate verification upfront when
+        connecting to the remote endpoint. This is on purpose, since this is
+        just a test procedure to get TLS and certificate information.
+
+        Certificate validity check is already done at OnionprobeHTTP.query_http().
 
         :type  endpoint: str
         :param endpoint: The endpoint name from the 'endpoints' instance config.
@@ -44,104 +53,218 @@ class OnionprobeTLS:
         :type  attempt: int
         :param attempt: The current attempt used to determine the maximum number of retries.
 
+        :rtype: bool
+        :return: True if the connection succeeded.
+                 False on error.
+
         """
 
-        tor_address            = self.get_config('tor_address')
-        socks_port             = self.get_config('socks_port')
-        timeout                = self.get_config('http_connect_timeout')
-        port                   = int(config['port']) if 'port' in config else 443
-        context                = ssl.create_default_context()
-        context.check_hostname = False
+        tor_address = self.get_config('tor_address')
+        socks_port  = self.get_config('socks_port')
+        timeout     = self.get_config('tls_connect_timeout')
+        port        = int(config['port']) if 'port' in config else 443
 
-        self.log('Retrieving X.509 certificate from {} on port {}...'.format(config['address'], port))
+        # Approach to use when always checking the certificate
+        #context                = ssl.create_default_context()
+        #context.check_hostname = True
+        #context.verify_mode    = ssl.CERT_REQUIRED
+        #valid_cert             = 1
+
+        # Approach to use to retrieve whichever certificate, no matter whether it's valid or not
+        context                = ssl.SSLContext()
+        context.check_hostname = False
+        context.verify_mode    = ssl.CERT_NONE
+        valid_cert             = 1
+
+        # Metric labels
+        labels = {
+                'name'    : endpoint,
+                'address' : config['address'],
+                'port'    : config['port'],
+                }
 
         try:
+            self.log('Trying to do a TLS connection to {} on port {} (attempt {})...'.format(
+                config['address'], port, attempt))
+
             with socks.create_connection(
                     (config['address'], port),
                     timeout=timeout, proxy_type=socks.SOCKS5,
                     proxy_addr=tor_address, proxy_port=socks_port, proxy_rdns=True) as sock:
                 with context.wrap_socket(sock, server_hostname=config['address']) as tls:
-                    cert_info = tls.getpeercert()
-                    result    = True
+                    result = True
 
-                    print(tls.version())
-                    print(cert_info)
+                    if self.get_config('get_certificate_info'):
+                        cert_result = self.get_certificate(endpoint, config, tls)
+
+                    self.info_metric('onion_service_tls_info', {
+                        'version'    : tls.version(),
+                        'compression': tls.compression(),
+                        'cipher'     : tls.cipher(),
+                        'stats'      : context.session_stats(),
+                        'alpn'       : tls.selected_alpn_protocol(),
+                        'npn'        : tls.selected_npn_protocol(),
+                        },
+                        labels)
 
         except ssl.SSLZeroReturnError as e:
-            result  = False
-            error   = e.reason
+            result    = False
+            error     = e.reason
+            exception = 'ssl_zero_return_error'
 
             self.log(e, 'error')
 
         except ssl.SSLWantReadError as e:
-            result  = False
-            error   = e.reason
+            result    = False
+            error     = e.reason
+            exception = 'ssl_want_read_error'
 
             self.log(e, 'error')
 
         except ssl.SSLWantWriteError as e:
-            result  = False
-            error   = e.reason
+            result    = False
+            error     = e.reason
+            exception = 'ssl_want_write_error'
 
             self.log(e, 'error')
 
         except ssl.SSLSyscallError as e:
-            result  = False
-            error   = e.reason
+            result    = False
+            error     = e.reason
+            exception = 'ssl_syscall_error'
 
             self.log(e, 'error')
 
         except ssl.SSLEOFError as e:
-            result  = False
-            error   = e.reason
+            result    = False
+            error     = e.reason
+            exception = 'ssl_eof_error'
 
             self.log(e, 'error')
 
         except ssl.SSLCertVerificationError as e:
-            result  = False
-            error   = e.reason
+            result     = False
+            error      = e.reason
+            exception  = 'ssl_cert_verification_error'
+            valid_cert = 0
 
             self.log(e, 'error')
 
         except ssl.CertificateError as e:
-            result  = False
-            error   = e.reason
+            result    = False
+            error     = e.reason
+            exception = 'ssl_certificate_error'
 
             self.log(e, 'error')
 
         except ssl.SSLError as e:
-            result  = False
-            error   = e.reason
+            result    = False
+            error     = e.reason
+            exception = 'ssl_error'
 
             self.log(e, 'error')
 
         except socks.GeneralProxyError as e:
-            result = False
-            error  = e.socket.err
+            result    = False
+            error     = e.socket.err
+            exception = 'general_proxy_error'
 
             self.log(e, 'error')
 
         except socks.SOCKS5AuthError as e:
-            result = False
-            error  = e.socket.err
+            result    = False
+            error     = e.socket.err
+            exception = 'socks5_auth_error'
 
             self.log(e, 'error')
 
         except socks.SOCKS5Error as e:
-            result = False
-            error  = e.socket.err
+            result    = False
+            error     = e.socket.err
+            exception = 'socks5_error'
 
             self.log(e, 'error')
 
         except socks.HTTPError as e:
-            result = False
-            error  = e.socket.err
+            result    = False
+            error     = e.socket.err
+            exception = 'http_error'
 
             self.log(e, 'error')
 
         except Exception as e:
-            result = False
+            result    = False
+            exception = 'general_error'
 
             self.log(e, 'error')
 
-        return result
+        finally:
+            reachable = 0 if result is False else 1
+
+            if result is False:
+                retries = self.get_config('tls_connect_max_retries')
+
+                # Try again until max retries is reached
+                if attempt <= retries:
+                    return self.query_tls(endpoint, config, attempt + 1)
+
+            return result
+
+    def get_certificate(self, endpoint, config, tls):
+        """
+        Get the certificate information from a TLS connection.
+
+        :type  endpoint: str
+        :param endpoint: The endpoint name from the 'endpoints' instance config.
+
+        :type  config: dict
+        :param config: Endpoint configuration
+
+        :type  tls: ssl.SSLSocket
+        :param tls: The TLS socket connection to the endpoint.
+
+        :rtype: bool
+        :return: True if the connection succeeded.
+                 False on error.
+
+        """
+
+        try:
+            # We can't rely on ssl.getpeercert() if the certificate wasn't validated
+            #cert_info = tls.getpeercert()
+
+            result           = True
+            der_cert         = tls.getpeercert(binary_form=True)
+            pem_cert         = ssl.DER_cert_to_PEM_cert(der_cert)
+            cert             = load_pem_x509_certificate(bytes(pem_cert, 'utf-8'))
+            cert_fp_sha1     = cert.fingerprint(hashes.SHA1()).hex(':').upper()
+            cert_fp_sha256   = cert.fingerprint(hashes.SHA256()).hex(':').upper()
+            not_valid_before = cert.not_valid_before.timestamp()
+            not_valid_after  = cert.not_valid_after.timestamp()
+
+            # Metric labels
+            labels = {
+                    'name'    : endpoint,
+                    'address' : config['address'],
+                    'port'    : config['port'],
+                    }
+
+            self.info_metric('onion_service_certificate_info', {
+                'version'           : str(cert.version),
+                'serial_number'     : cert.serial_number,
+                'fingerprint_sha1'  : cert_fp_sha1,
+                'fingerprint_sha256': cert_fp_sha256,
+                },
+                labels)
+
+            self.set_metric('onion_service_certificate_not_valid_before', not_valid_before, labels)
+            self.set_metric('onion_service_certificate_not_valid_after',  not_valid_after,  labels)
+
+        except Exception as e:
+            result    = False
+            exception = 'general_error'
+
+            self.log(e, 'error')
+
+        finally:
+            return result